chore: move DefinitelyTyped types to Network.d.ts
